feat: move dev tools to workspace menu
