+++ src/plugins/dev_tools/public/dev_tools_modal.scss
